feat: get positions api
