+++ src/events/getHistoricInvests.event.js
@@ -38,7 +38,7 @@ async function processHistoricInvestEvents() {
       
           const blockTimeStamp = new Date(block.timestamp * 1000);
           
-          await axios.post(`${baseURL}/api/v1/topic/invest`, {
+          await axios.post(`${baseURL}/api/v1/invest/createInvest`, {
             investor,
             topicId: topicId.toString(),
             amount: decimalAmount,
